+++ src/wrappers/mj_visualization.rs
@@ -84,7 +84,7 @@ impl MjvPerturb {
 
     pub fn apply<M: Deref<Target = MjModel>>(&mut self, model: &MjModel, data: &mut MjData<M>) {
         unsafe {
-            mju_zero(data.ffi_mut().xfrc_applied, 6 * model.ffi().nbody);
+            mju_zero(data.ffi_mut().xfrc_applied, 6 * model.ffi().nbody as i32);
             mjv_applyPerturbPose(model.ffi(), data.ffi_mut(), self, 0);
             mjv_applyPerturbForce(model.ffi(), data.ffi_mut(), self);
         }
